LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssColorValue-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: CSSHSL

Harness Error (FAIL), message = ReferenceError: Can't find variable: CSSHSL
